railway-backend/requirements.txt
openai>=1.0.0
requests>=2.25.0
tweepy>=4.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pandas>=1.3.0
python-dotenv>=0.19.0
